Switchless Ocall Worker Waiting Strategy Improvement
Each host worker checks for a message in its call_arg slot.
If there is a message, the message is immediately removed from
the slot and processed.

If there is no message, then a spin counter is incremented.
If the spin counter reaches a particular threshold, then it means
that the thread has been spinning for sometime waiting for messages
and now can be put to sleep.

A simple strategy of putting the thread to sleep is to sleep for
a predetermined amount of time say 1 microsecond. This has the
benefit of letting other threads run; however it does not give
optimal performance since the worker threads keep waking up,
spinning and going back to sleep for ever, consuming enough CPU
to affect normal ocall performance.

A better strategy is to use a futex to wait.
Each thread has an int32_t event field, initially zero.

When a host thread has spun till the threshold,
    if atomically sets the value of event to 0 if it was 1

    if the previous value (returned from atomic operation) was 0,
    it means that there has been no wake notifications from the enclave,
    and therefore the host thread does a futex_wait while event's value
    is still zero.

    if the previous value (returned from atomic operation) was 1,
    if means that the enclave has just posted a message and therefore
    the worker thread continues to fetch the message instead of sleeping.

When an enclave thread wants to post a switchless ocall message,
   it first finds a host worker thread with a free slot
       The host worker thread could be active or doing a futex_wait

       The enclave thread atomically sets the value of the host thread's
       event to 1 if it was 0

       if the previous value(returned from atomic operation) was 0,
       it means that the host thread is doing a futex_wait and therefore
       an core ocall (OE_OCALL_WAKE_HOST_WORKER) is made to futex_wake the thread.
       The number of such ocalls will be less in an application that makes switchless
       calls with enough frequency. (If switchless calls are rare, the application may
       not benefit from switchless calls).

       If the previous value(returned from atomic operation) was 1,
       it means that we have already woken up the host and therefore don't need to
       wake up the host again.

   if no host worker thread with free slot is found, the post fails and
   subsquently a normal ocall is made

On the enclave side, an atomic compare and exchange operation suffices to determine
if the host must be woken up via a core ocall or not.

On the host side, in linux futex syscalls (__NR_futex FUTEX_WAIT_PRIVATE and FUTEX_WAKE_PRIVATE)
are used whereas on Windows WaitOnAddress and WakeByAddressSingle are used.